user profile & upload

+++ src/controllers/videosController.js
@@ -1,5 +1,6 @@
 //Fake video
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async (req, res) => {
     const videos = await Video.find({}).sort({createdAt: "desc"});
@@ -9,7 +10,7 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
     const {id} = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
     if(!video){
         return res.render("watch",{pageTitle: video.title, video});
     }
@@ -18,19 +19,31 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
     const {id} = req.params;
+    const {
+        user: { _id },
+    } = req.session;
     const video = await Video.findById(id);
     if(!video){
         return res.status(404).render("404",{pageTitle:"Video not found."});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     return res.render("edit",{pageTitle:`Edit: ${video.title}`, video});
 };
 
 export const postEdit = async (req, res) => {
+    const {
+        user: { _id },
+    } = req.session;
     const {id} = req.params;
     const {title, description, hashtags} = req.body;
-    const video = await Video.exists({_id: id});
+    const video = await Video.findById(id);
     if(!video){
         return res.status(404).render("404",{pageTitle: "Video not found."});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -45,16 +58,23 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+    const {
+        user: { _id },
+    } = req.session;
     const {path: fileUrl} = req.file;
     const {title, description, hashtags} = req.body;
     try{
-        console.log(title);
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
             fileUrl,
+            owner: _id,
             hashtags: Video.formatHashtags(hashtags),
-        });        
+        });
+        //  throw "sssss";
+        const user = await User.findById(_id); 
+        user.videos.push(newVideo._id);
+        user.save(); 
         return res.redirect("/");
     }catch(error){
         return res.status(400).render("upload",{
@@ -66,6 +86,16 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const {id} = req.params;
+    const {
+        user: { _id },
+    } = req.session;
+    const video = await Video.findById(id);
+    if(!video){
+        return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 }
